Former-commit-id: b2cd17f68be54289962f778c7a65109b0ae5f8f3

+++ fingerprint/src/main/java/com/simprints/fingerprint/scanner/domain/ScannerVersionInformation.kt
@@ -1,5 +1,5 @@
 package com.simprints.fingerprint.scanner.domain
 
 class ScannerVersionInformation(val veroVersion: Int,
-                                val firmwareVersion: Int,
-                                val un20Version: Int)
+                                val firmwareVersion: Long,
+                                val un20Version: Long)
